refactor: 객체명 수정
